refactor: altera path repository e dtos
