+++ src/main/java/com/friendlyblood/api/domain/services/UserService.java
@@ -1,7 +1,8 @@
 package com.friendlyblood.api.domain.services;
 
 import com.friendlyblood.api.domain.models.User;
-import com.friendlyblood.api.domain.repositories.UserRepository;
+import com.friendlyblood.api.repositories.UserRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
